+++ packages/snap-preact-components/src/components/Molecules/Carousel/Carousel.tsx
@@ -6,7 +6,7 @@ import { jsx, css } from '@emotion/react';
 import classnames from 'classnames';
 import { observer } from 'mobx-react-lite';
 import deepmerge from 'deepmerge';
-import SwiperCore, { Pagination, Navigation } from 'swiper/core';
+import SwiperCore, { Pagination, Navigation, A11y } from 'swiper/core';
 import { SwiperOptions } from 'swiper';
 import { Icon, IconProps } from '../../Atoms/Icon/Icon';
 import { Swiper, SwiperSlide } from 'swiper/react';
@@ -24,6 +24,11 @@ const CSS = {
 			margin: 0,
 			padding: 0,
 			overflow: 'hidden',
+
+			'.swiper-notification': {
+				position: 'absolute',
+				left: '100000000000000px',
+			},
 
 			'&.ss__carousel-vertical': {
 				flexDirection: 'column',
@@ -246,7 +251,7 @@ export const Carousel = observer((properties: CarouselProps): JSX.Element => {
 		},
 	};
 
-	const swiperModulesUnfiltered = modules ? [Navigation, Pagination].concat(modules!) : [Navigation, Pagination];
+	const swiperModulesUnfiltered = modules ? [Navigation, Pagination, A11y].concat(modules!) : [Navigation, Pagination, A11y];
 	//remove any duplicates, in case user passes in Navigation or Pagination
 	const swiperModules = swiperModulesUnfiltered.filter((module, pos) => swiperModulesUnfiltered.indexOf(module) === pos);