+++ packages/snap-preact-components/src/components/Organisms/Autocomplete/Autocomplete.tsx
@@ -208,11 +208,15 @@ export const Autocomplete = observer((properties: AutocompleteProps): JSX.Elemen
 
 	let delayTimeout: number;
 	const delayTime = 333;
+
 	const valueProps = {
-		onMouseEnter: (e: React.MouseEvent<HTMLAnchorElement>) => {
+		onMouseEnter: (e: React.MouseEvent<HTMLAnchorElement>, activeTerm: any) => {
 			clearTimeout(delayTimeout);
 			delayTimeout = window.setTimeout(() => {
 				(e.target as HTMLAnchorElement).focus();
+				if (activeTerm && activeTerm.preview) {
+					activeTerm.preview();
+				}
 			}, delayTime);
 		},
 		onMouseLeave: () => {
@@ -369,7 +373,7 @@ export const Autocomplete = observer((properties: AutocompleteProps): JSX.Elemen
 		},
 	};
 
-	const { search, terms, trending, results, merchandising, pagination, loaded, filters, facets, state } = controller.store;
+	const { search, terms, trending, results, merchandising, pagination, loaded, filters, facets, state, loading } = controller.store;
 	const history = controller.store.history || [];
 
 	// you can pass in a selector or the actual input element,
@@ -386,29 +390,30 @@ export const Autocomplete = observer((properties: AutocompleteProps): JSX.Elemen
 		Boolean(input === state.focusedInput) &&
 		(terms.length > 0 || trending?.length > 0 || history?.length > 0 || (state.input && controller.store.loaded));
 
+	const trendingActive = trending?.filter((term) => term.active).pop();
+	const historyActive = history?.filter((term) => term.active).pop();
+
 	let showTrending = false;
-	if (trending?.length && (retainTrending || (!results.length && !state.input))) {
-		showTrending = true;
-	} else if (trending?.length && !terms.length) {
-		// has results and trending -> show trending terms while term load
+	if (trending?.length && ((retainTrending && controller.store.loaded) || (!results.length && !state.input))) {
 		showTrending = true;
 	}
 
 	let showHistory = false;
-	if (history?.length && (retainHistory || (!results.length && !state.input))) {
-		showHistory = true;
-	} else if (history?.length && !terms.length) {
-		// has results and trending -> show trending terms while term load
+	if (history?.length && ((retainHistory && controller.store.loaded) || (!results.length && !state.input))) {
 		showHistory = true;
 	}
 
+	if (!state.input && (historyActive || trendingActive)) {
+		showHistory = true;
+		showTrending = true;
+	}
+
 	const facetsToShow = facets.length ? facets.filter((facet) => facet.display !== FacetDisplay.SLIDER) : [];
-	const onlyTerms = (trending?.length || history.length) && !loaded;
+	const onlyTerms = (trending?.length || history.length) && !loaded && !loading;
 
 	// results logic
 	let showResults = Boolean(results.length > 0 || Object.keys(merchandising.content).length > 0 || search?.query?.string);
-	const trendingActive = trending?.filter((term) => term.active).pop();
-	const historyActive = history?.filter((term) => term.active).pop();
+
 	if ((hideTrending && trendingActive) || (hideHistory && historyActive)) {
 		showResults = false;
 	}
@@ -465,8 +470,8 @@ export const Autocomplete = observer((properties: AutocompleteProps): JSX.Elemen
 												<h5>{termsTitle}</h5>
 											</div>
 										) : null}
-										<div className="ss__autocomplete__terms__options">
-											{terms.map((term) => (
+										<div className="ss__autocomplete__terms__options" role={'list'} aria-label={termsTitle}>
+											{terms.map((term, idx) => (
 												<div
 													className={classnames('ss__autocomplete__terms__option', {
 														'ss__autocomplete__terms__option--active': term.active,
@@ -476,7 +481,9 @@ export const Autocomplete = observer((properties: AutocompleteProps): JSX.Elemen
 														onClick={(e: React.MouseEvent<HTMLAnchorElement, MouseEvent>) => termClickEvent(e)}
 														href={term.url.href}
 														{...valueProps}
-														onFocus={() => term.preview()}
+														onMouseEnter={(e) => valueProps.onMouseEnter(e, term)}
+														role="link"
+														aria-label={`item ${idx + 1} of ${terms.length}, ${term.value}`}
 													>
 														{emIfy(term.value, state.input || '')}
 													</a>
@@ -493,8 +500,8 @@ export const Autocomplete = observer((properties: AutocompleteProps): JSX.Elemen
 												<h5>{trendingTitle}</h5>
 											</div>
 										) : null}
-										<div className="ss__autocomplete__terms__options">
-											{trending.map((term) => (
+										<div className="ss__autocomplete__terms__options" role={'list'} aria-label={trendingTitle}>
+											{trending.map((term, idx) => (
 												<div
 													className={classnames('ss__autocomplete__terms__option', {
 														'ss__autocomplete__terms__option--active': term.active,
@@ -504,7 +511,9 @@ export const Autocomplete = observer((properties: AutocompleteProps): JSX.Elemen
 														onClick={(e: React.MouseEvent<HTMLAnchorElement, MouseEvent>) => termClickEvent(e)}
 														href={term.url.href}
 														{...valueProps}
-														onFocus={() => term.preview()}
+														onMouseEnter={(e) => valueProps.onMouseEnter(e, term)}
+														role="link"
+														aria-label={`item ${idx + 1} of ${trending.length}, ${term.value}`}
 													>
 														{emIfy(term.value, state.input || '')}
 													</a>
@@ -521,8 +530,8 @@ export const Autocomplete = observer((properties: AutocompleteProps): JSX.Elemen
 												<h5>{historyTitle}</h5>
 											</div>
 										) : null}
-										<div className="ss__autocomplete__terms__options">
-											{history.map((term) => (
+										<div className="ss__autocomplete__terms__options" role={'list'} aria-label={historyTitle}>
+											{history.map((term, idx) => (
 												<div
 													className={classnames('ss__autocomplete__terms__option', {
 														'ss__autocomplete__terms__option--active': term.active,
@@ -532,7 +541,9 @@ export const Autocomplete = observer((properties: AutocompleteProps): JSX.Elemen
 														onClick={(e: React.MouseEvent<HTMLAnchorElement, MouseEvent>) => termClickEvent(e)}
 														href={term.url.href}
 														{...valueProps}
-														onFocus={() => term.preview()}
+														onMouseEnter={() => term.preview()}
+														role="link"
+														aria-label={`item ${idx + 1} of ${history.length}, ${term.value}`}
 													>
 														{emIfy(term.value, state.input || '')}
 													</a>