+++ packages/snap-preact-components/src/hooks/index.ts
@@ -3,3 +3,4 @@ export * from './useConstructor';
 export * from './useMediaQuery';
 export * from './useIntersection';
 export * from './useDisplaySettings';
+export * from './useA11y';